+++ wind-core/src/main/java/io/github/ramerf/wind/core/service/UpdateService.java
@@ -2,9 +2,11 @@ package io.github.ramerf.wind.core.service;
 
 import io.github.ramerf.wind.core.condition.ICondition;
 import io.github.ramerf.wind.core.entity.pojo.AbstractEntityPoJo;
+import io.github.ramerf.wind.core.entity.request.AbstractEntityRequest;
 import io.github.ramerf.wind.core.entity.response.ResultCode;
 import io.github.ramerf.wind.core.exception.CommonException;
 import io.github.ramerf.wind.core.util.CollectionUtils;
+import java.util.Collection;
 import java.util.List;
 import java.util.function.Consumer;
 import javax.annotation.Nonnull;
@@ -54,6 +56,84 @@ public interface UpdateService<T extends AbstractEntityPoJo> extends InterServic
       return 0;
     }
     return CollectionUtils.isEmpty(ts) ? 0 : getUpdate().createBatch(ts);
+  }
+
+  /**
+   * 更新不为null的字段..
+   *
+   * @param t the t
+   * @return the t
+   * @throws RuntimeException the runtime exception
+   */
+  default int update(T t) throws RuntimeException {
+    return getUpdate().update(t, false);
+  }
+
+  /**
+   * 注意: 该方法只会更新不为null的字段.
+   *
+   * @param t the t
+   * @param consumer 更新条件
+   * @return the 更新记录数
+   * @throws RuntimeException the runtime exception
+   */
+  default int update(T t, Consumer<ICondition<T>> consumer) throws RuntimeException {
+    return getUpdate().where(consumer).update(t, false);
+  }
+
+  /**
+   * 批量更新.<br>
+   *
+   * @param ts 要更新的数据集
+   * @param includeNullProperties the include null properties
+   * @return {@link AbstractEntityPoJo}
+   * @throws RuntimeException sqlException
+   */
+  @Transactional(rollbackFor = Exception.class)
+  default int updateBatch(final List<T> ts, final String... includeNullProperties)
+      throws RuntimeException {
+    return getUpdate().updateBatch(ts, false);
+  }
+
+  /**
+   * 批量更新.<br>
+   * 注意: 该方法会将ts中的值完全覆盖对应的数据.<br>
+   *
+   * @param <R> the type parameter
+   * @param ts 要更新的数据集
+   * @param includeNullProperties the include null properties
+   * @return {@link AbstractEntityPoJo}
+   * @throws RuntimeException sqlException
+   */
+  @Transactional(rollbackFor = Exception.class)
+  default <R extends AbstractEntityRequest<?>> int updateBatchRequest(
+      List<R> ts, final String... includeNullProperties) throws RuntimeException {
+    return getUpdate().updateBatchRequest(ts, false);
+  }
+
+  /**
+   * Delete.
+   *
+   * @param id the id
+   * @throws RuntimeException the runtime exception
+   */
+  default void delete(final long id) throws RuntimeException {
+    getUpdate().where(condition -> condition.eq(AbstractEntityPoJo::setId, id)).delete();
+  }
+
+  /**
+   * 如果删除数量不等于id的大小,将执行失败.
+   *
+   * @param ids the ids
+   * @return 删除记录数 int
+   * @throws RuntimeException the runtime exception
+   */
+  @Transactional(rollbackFor = Exception.class)
+  default int deleteBatch(final Collection<Long> ids) throws RuntimeException {
+    if (CollectionUtils.isEmpty(ids)) {
+      return 0;
+    }
+    return getUpdate().where(condition -> condition.in(AbstractEntityPoJo::setId, ids)).delete();
   }
 
   /**
